+++ models/actions/task_test.go
@@ -5,7 +5,6 @@ package actions
 
 import (
 	"strings"
-	"sync"
 	"testing"
 
 	runnerv1 "gitea.dev/actions-proto-go/runner/v1"
@@ -306,105 +305,6 @@ func TestStopTaskCancellingFallsBackForMissingRunner(t *testing.T) {
 	jobAfterStop := unittest.AssertExistsAndLoadBean(t, &ActionRunJob{ID: job.ID})
 	assert.Equal(t, StatusCancelled, jobAfterStop.Status)
 	assert.NotZero(t, jobAfterStop.Stopped)
-}
-
-// minimalWorkflowPayload returns the minimal YAML for a single-job workflow with no steps.
-func minimalWorkflowPayload(jobID string) []byte {
-	return []byte("on: push\njobs:\n  " + jobID + ":\n    runs-on: ubuntu-latest\n")
-}
-
-// TestCreateTaskForRunnerConcurrentClaim verifies that when multiple runners
-// poll simultaneously and all initially see the same first waiting job,
-// each runner claims a distinct job rather than all but one being left
-// empty-handed. This is the regression test for the race condition where
-// runners losing the optimistic-lock on job #1 would receive latestVersion
-// and never retry the remaining 49+ jobs.
-func TestCreateTaskForRunnerConcurrentClaim(t *testing.T) {
-	require.NoError(t, unittest.PrepareTestDatabase())
-
-	const numJobs = 3
-
-	run := &ActionRun{
-		Title:         "concurrent-claim-test-run",
-		RepoID:        1,
-		OwnerID:       2,
-		WorkflowID:    "test.yaml",
-		Index:         9901,
-		TriggerUserID: 2,
-		Ref:           "refs/heads/main",
-		CommitSHA:     "c2d72f548424103f01ee1dc02889c1e2bff816b0",
-		Event:         "push",
-		TriggerEvent:  "push",
-		Status:        StatusWaiting,
-	}
-	require.NoError(t, db.Insert(t.Context(), run))
-
-	jobs := make([]*ActionRunJob, numJobs)
-	for i := range numJobs {
-		jobID := "concurrent-job-" + string(rune('a'+i))
-		jobs[i] = &ActionRunJob{
-			RunID:           run.ID,
-			RepoID:          run.RepoID,
-			OwnerID:         run.OwnerID,
-			CommitSHA:       run.CommitSHA,
-			Name:            jobID,
-			Attempt:         1,
-			JobID:           jobID,
-			Status:          StatusWaiting,
-			RunsOn:          []string{"ubuntu-latest"},
-			WorkflowPayload: minimalWorkflowPayload(jobID),
-		}
-		require.NoError(t, db.Insert(t.Context(), jobs[i]))
-	}
-
-	runners := make([]*ActionRunner, numJobs)
-	for i := range numJobs {
-		r := &ActionRunner{
-			UUID:        "concurrent-runner-uuid-" + string(rune('a'+i)),
-			Name:        "concurrent-runner-" + string(rune('a'+i)),
-			AgentLabels: []string{"ubuntu-latest"},
-		}
-		r.GenerateAndFillToken()
-		runners[i] = r
-		require.NoError(t, db.Insert(t.Context(), runners[i]))
-	}
-
-	// Simulate the burst: all runners call CreateTaskForRunner concurrently,
-	// as happens when all see the same stale tasksVersion simultaneously.
-	type result struct {
-		task *ActionTask
-		ok   bool
-		err  error
-	}
-	results := make([]result, numJobs)
-	var wg sync.WaitGroup
-	for i := range numJobs {
-		wg.Add(1)
-		go func(i int) {
-			defer wg.Done()
-			task, ok, err := CreateTaskForRunner(t.Context(), runners[i])
-			results[i] = result{task, ok, err}
-		}(i)
-	}
-	wg.Wait()
-
-	// Every runner must have received a task without error.
-	claimedJobIDs := make(map[int64]bool)
-	for i, r := range results {
-		require.NoError(t, r.err, "runner %d got an unexpected error", i)
-		require.True(t, r.ok, "runner %d did not get a task even though free jobs exist", i)
-		require.NotNil(t, r.task)
-		assert.False(t, claimedJobIDs[r.task.JobID], "job %d was claimed by more than one runner", r.task.JobID)
-		claimedJobIDs[r.task.JobID] = true
-	}
-	assert.Len(t, claimedJobIDs, numJobs, "expected %d distinct jobs to be claimed", numJobs)
-
-	// All jobs must now be running with a task assigned.
-	for _, j := range jobs {
-		updated := unittest.AssertExistsAndLoadBean(t, &ActionRunJob{ID: j.ID})
-		assert.Equal(t, StatusRunning, updated.Status)
-		assert.NotZero(t, updated.TaskID)
-	}
 }
 
 // TestReleaseTaskForRunner verifies that releasing a freshly-claimed task returns
